test(daemon): split plist_argv0 tests — keep portable Plist() tests on all OSes
Refines fe3c054: instead of darwin-gating the whole file, split it. The two
Plist()-generation tests (TestPlistProdSplitsProgramFromArgv0,
TestPlistTestModeKeepsLegacyArgv) exercise pure cross-platform string building
and now stay ungated so linux CI keeps covering the Program/argv0-split
behaviour. Only the two parsePlist read-back tests (parsePlist lives in
ctl_darwin.go) move to a new //go:build darwin file. Addresses review MEDIUM:
the whole-file gate silently dropped linux coverage for portable logic.
